gdrive: remove old drive implementation
