+++ app/build.gradle
@@ -84,10 +84,6 @@ dependencies {
     implementation 'androidx.appcompat:appcompat:1.1.0'
     implementation 'androidx.multidex:multidex:2.0.1'
 
-    implementation "com.google.android.gms:play-services-base:11.0.2"
-    implementation "com.google.android.gms:play-services-drive:11.0.2"
-    implementation "com.google.android.gms:play-services-plus:11.0.2"
-
     annotationProcessor "org.androidannotations:androidannotations:$AAVersion"
     implementation "org.androidannotations:androidannotations-api:$AAVersion"
 
@@ -97,7 +93,7 @@ dependencies {
     implementation 'com.dropbox.core:dropbox-core-sdk:3.0.8'
     implementation 'com.squareup.okhttp3:okhttp:3.12.6'
     implementation 'com.squareup.okio:okio:1.17.5'
-    implementation 'org.greenrobot:eventbus:3.1.1'
+    implementation 'org.greenrobot:eventbus:3.3.1'
     implementation 'com.wdullaer:materialdatetimepicker:3.6.4'
     implementation 'commons-io:commons-io:2.5'
     implementation 'com.mtramin:rxfingerprint:2.2.1'
@@ -106,6 +102,11 @@ dependencies {
     implementation 'org.reactivestreams:reactive-streams:1.0.2'
     implementation 'com.mlsdev.rximagepicker:library:2.1.5'
     implementation 'com.github.bumptech.glide:glide:4.10.0'
+
+    implementation 'com.google.android.gms:play-services-auth:16.0.1'
+    implementation 'com.google.http-client:google-http-client-gson:1.26.0'
+    implementation 'com.google.api-client:google-api-client-android:1.26.0'
+    implementation 'com.google.apis:google-api-services-drive:v3-rev136-1.25.0'
 
     implementation fileTree(include: '**/*.jar', dir: 'libs')
