+++ app/src/main/java/ru/orangesoftware/financisto/export/Export.java
@@ -29,8 +29,8 @@ import java.util.zip.GZIPOutputStream;
 
 import ru.orangesoftware.financisto.R;
 import ru.orangesoftware.financisto.activity.RequestPermission;
-import ru.orangesoftware.financisto.export.drive.GoogleDriveClient;
-import ru.orangesoftware.financisto.export.drive.GoogleDriveClient_;
+import ru.orangesoftware.financisto.export.drive.GoogleDriveClientV3;
+import ru.orangesoftware.financisto.export.drive.GoogleDriveClientV3_;
 import ru.orangesoftware.financisto.export.dropbox.Dropbox;
 import ru.orangesoftware.financisto.utils.MyPreferences;
 
@@ -133,7 +133,7 @@ public abstract class Export {
 
     public static void uploadBackupFileToGoogleDrive(Context context, String backupFileName) throws Exception {
         File file = getBackupFile(context, backupFileName);
-        GoogleDriveClient driveClient = GoogleDriveClient_.getInstance_(context);
+        GoogleDriveClientV3 driveClient = GoogleDriveClientV3_.getInstance_(context);
         driveClient.uploadFile(file);
     }
