--- app/src/main/java/ru/orangesoftware/financisto/export/drive/DoDriveBackup.java
@@ -1,4 +1,0 @@
-package ru.orangesoftware.financisto.export.drive;
-
-public class DoDriveBackup {
-}
