+++ app/src/main/java/ru/orangesoftware/financisto/export/drive/DoDriveBackupEvent.java
@@ -1,0 +1,6 @@
+package ru.orangesoftware.financisto.export.drive;
+
+public class DoDriveBackupEvent {
+    public DoDriveBackupEvent() {
+    }
+}
