--- app/src/main/java/ru/orangesoftware/financisto/export/drive/DoDriveListFiles.java
@@ -1,4 +1,0 @@
-package ru.orangesoftware.financisto.export.drive;
-
-public class DoDriveListFiles {
-}
