--- app/src/main/java/ru/orangesoftware/financisto/export/drive/DoDriveRestore.java
@@ -1,11 +1,0 @@
-package ru.orangesoftware.financisto.export.drive;
-
-public class DoDriveRestore {
-
-    public final DriveFileInfo selectedDriveFile;
-
-    public DoDriveRestore(DriveFileInfo selectedDriveFile) {
-        this.selectedDriveFile = selectedDriveFile;
-    }
-
-}
