+++ app/src/main/java/ru/orangesoftware/financisto/export/drive/DoDriveRestoreEvent.java
@@ -1,0 +1,6 @@
+package ru.orangesoftware.financisto.export.drive;
+
+public class DoDriveRestoreEvent {
+    public DoDriveRestoreEvent() {
+    }
+}
