--- app/src/main/java/ru/orangesoftware/financisto/export/drive/DriveBackupError.java
@@ -1,11 +1,0 @@
-package ru.orangesoftware.financisto.export.drive;
-
-public class DriveBackupError {
-
-    public final String message;
-
-    public DriveBackupError(String message) {
-        this.message = message;
-    }
-
-}
