--- app/src/main/java/ru/orangesoftware/financisto/export/drive/DriveBackupSuccess.java
@@ -1,11 +1,0 @@
-package ru.orangesoftware.financisto.export.drive;
-
-public class DriveBackupSuccess {
-
-    public final String fileName;
-
-    public DriveBackupSuccess(String fileName) {
-        this.fileName = fileName;
-    }
-
-}
