--- app/src/main/java/ru/orangesoftware/financisto/export/drive/DriveConnectionFailed.java
@@ -1,13 +1,0 @@
-package ru.orangesoftware.financisto.export.drive;
-
-import com.google.android.gms.common.ConnectionResult;
-
-public class DriveConnectionFailed {
-
-    public final ConnectionResult connectionResult;
-
-    public DriveConnectionFailed(ConnectionResult connectionResult) {
-        this.connectionResult = connectionResult;
-    }
-
-}
