--- app/src/main/java/ru/orangesoftware/financisto/export/drive/DriveFileList.java
@@ -1,13 +1,0 @@
-package ru.orangesoftware.financisto.export.drive;
-
-import java.util.List;
-
-public class DriveFileList {
-
-    public final List<DriveFileInfo> files;
-
-    public DriveFileList(List<DriveFileInfo> files) {
-        this.files = files;
-    }
-
-}
