--- app/src/main/java/ru/orangesoftware/financisto/export/drive/DriveRestoreSuccess.java
@@ -1,4 +1,0 @@
-package ru.orangesoftware.financisto.export.drive;
-
-public class DriveRestoreSuccess {
-}
